Packaging powered by make
Thanks to narc0tiq for the factorio-mod-makefile project, which provided the
starting point for this packaging system. It has proven itself over the last
two releases of Warehousing, and is now ready to be included in the codebase.

(Actually, it was meant to be committed after the 0.0.13 release, months ago,
but I got distracted by other projects and forgot to do it.)