Base logistics warehousing research cost on Logistic System cost
Players who use mods to reduce the cost of Logistic System will appreciate this.

Used data-final-fixes.lua to copy the tech cost as late as possible, but it's possible
that some mod which alters the cost of Logistic System in the base game will also use
the same load step. They can be made to load first by adding them as optional
dependencies of this mod.