Add changelog.txt with history going back to 0.0.13
Factorio including a changelog viewer for mods directly in the game is really
quite exciting. At some point I might go back and cobble together a version
history that goes all the way back to the first release, but I didn't want to
spend too long on this and delay 0.1.1 any more.