Add placeholder locale files for 18 new target languages
Enabled the most common languages on Crowdin. It added:

  - Arabic
  - Catalan
  - Chinese Simplified (this was requested in a discussion thread)
  - Czech
  - Danish
  - Dutch
  - Finnish
  - Greek
  - Hebrew
  - Italian
  - Norwegian
  - Portuguese (not Brazilian)
  - Romanian
  - Serbian (Cyrillic)
  - Swedish
  - Turkish
  - Ukranian
  - Vietnamese

While it would be preferable to only include locale files for languages
with available translations, Crowdin will insistently add empty files
anyway. There's no setting to disable this behavior, so we have to work
with it instead. Empty locale files are small, fortunately.

The advantage of having languages enabled with no content is that
translators can immediately get to work on Crowdin without having to try
and ping a project maintainer to add the language they want to work on.
Less common languages will still have this problem, but these additions
cover a huge percentage of the world's population.