Rebalance building storage capacities
Went with 50 slots/tile for the nice round numbers (450 & 1800).

Storage warehouse has not changed. Storage storehouse now stores exactly
one quarter the amount of items as the storage warehouse—which matches
the entities' relative footprints (3x3 is one quarter of 6x6).

Storage variants *may* still change if I can devise a good way of
migrating existing buildings. I only left them buffed compared to the
other building types because shrinking a building's inventory size leads
to vanishing items if an existing building is filled past the new limit.
A migration might be able to help with this, but I'm not yet sure how.
There isn't anywhere for the extra items to *go*, is the problem.

So, for now, storage variants will keep their buff. It's just a much
smaller buff (relative to the total size) than it used to be.