Tentative research cost rebalancing
Basic warehousing research time per-unit was boosted from 20 to 30
seconds, as a symbolic sacrifice in exchange for the much-higher
building capacities planned as part of this rebalancing.

Logistic warehousing is MUCH more expensive—250 cycles, up from 100 (or
150, taking into account the fact that Warehousing actually copies the
cost of "Logistic System" from the base game and overwrites its default
research cost with that). However, it no longer requires Production
Science Packs, because Factorio 0.17 (as of FFF 275) dropped that
requirement for the logistic system research.

This commit will probably need to be amended after 0.17 drops, due to
item names changing and/or recipes being different from what was
announced in FFF 275.

Random whitespace changes included because making a separate commit for
those seemed extremely pointless.