Overhaul crafting recipes for warehouse/storehouse buildings
Wooden chests are gone!

That was the primary goal of reworking these recipes, but some other
stuff happened along the way too.

Using both steel and iron plate seemed odd, so only steel is used now.
(This works nicely with the fact that steel processing is a prerequisite
for researching warehouse tech.)

Iron is still a component, though: The buildings use iron sticks for
internal bracing, now. Think of the truss-like structures you'd see on
the ceiling of a real warehouse, or the floor-to-ceiling supports a real
building has to have (whether hidden inside a wall or not). That's what
the iron sticks are for.

Stone brick amounts were adjusted to make sense for the footprint of
each building size while still being round numbers.

Logistic variants don't need circuits of any kind, any more. All the
smarts come from the logistic chest that gets tossed in. The extra
ingredients for logistic versions are to build the extra bits you might
notice on the roof: those vents etc. that get added on top. They're made
from iron sticks and steel plate, just like the regular buildings.

Like the rest of the whole rebalancing for Factorio 0.17, these tweaks
are subject to change.