Create "hero" image and thumbnail
Factorio 0.17 will display mod thumbnails in its GUI, using the
`thumbnail.png` file included in the mod package.

Since I needed a screenshot to create the thumbnail anyway, I decided to
also make a "hero" screenshot with all the buildings in the mod arranged
in a semi-plausible use case and put that picture in the Readme.