Update screenshot/thumbnail/demo-map for 0.17
Can't use Creative Mode any more, but I definitely can throw together
enough coal power to run the demo setup long enough to take a picture.

Updated the save file (and associated notes), the default hero image,
the hero image displayed in the Readme, and the thumbnail.

Also used an absolute URL to the hero-image in the Readme for
consistency (and so pasting the Readme into the Mod Portal should work).