Add localizable mod name/description
This is going to make Crowdin generate a TON of commits in the open l10n
PR, adding these to all the other locale files, but there isn't any way
around that. Crowdin's gonna do what Crowdin's gonna do.

It would be nice if the info.json file could pull these localized values
in, instead of needing hardcoded title/description attributes, but I
guess localization support will be added to the mod portal in a future
version (and for now, we just have to deal with the string duplication).