Consistent vector style in entity.lua
I could write some BS reason for doing this, but the real reason is
quite simple: I was looking for some insignificant change to make as an
excuse to push a new, non-merge commit to `master`, because Crowdin is
acting up and I suspect it has to do with the tip of `master` being a
commit with two parents, or something silly like that.

So yeah, this is really just a slightly more useful alternative to doing
a `git commit --allow-empty` and pushing that.