Update screenshot-map to load in Factorio 1.1
Leftover script/scenario things both in the ZIP and outside it caused
problems loading the file in 1.1. See https://forums.factorio.com/92500

On the plus side, the save file is now much smaller.

Since it was last touched in some version of Factorio 0.17.x, updating
the save required first downloading Factorio 1.0.0 to re-save the map in
a supported version (1.1 won't load anything older than 0.18.0) before
the script issue ever came up. It was indeed a Process to get this done.