Reorganize most mod files into `src/` subdirectory
Had to rewrite a good chunk of the Makefile to accommodate this, but the
end result is exactly as desired. I compared the new `make` result with
an old output folder renamed to `pkg/oldMakefile`:

    $ diff -u -r pkg/oldMakefile pkg/Warehousing_0.6.0
    Only in pkg/oldMakefile: .vscode
    Only in pkg/oldMakefile: CONTRIBUTING.md
    Only in pkg/oldMakefile: migrations

Works for me:

- `.vscode` never should be shipped
  (It also shouldn't be in Git, but I have a separate "housekeeping" PR
  for things like this.)
- The `migrations` do nothing, so they aren't moved to `src/`
  (My "housekeeping" PR will cover this too.)
- `CONTRIBUTING.md` doesn't make sense to include in the package

The only thing to look out for in the future is, if we want to use any
placeholders the `SED_EXPRS` handle (like `{{VERSION}}`), right now the
file in question will have to live in the `SOURCE_DIR` in order to match
the relevant rule. We aren't using replacements outside of `*.json` or
`*.lua`, so I just didn't bother implementing that in the rewrite.