Actions: Specify runner environment + lower Lua version
Factorio's internal Lua interpreter seems to be based on Lua 5.2.1. That
isn't guaranteed to stay the same, and installing a specific version of
the lua5.2 package would probably be overkill. But it seems smart to run
release builds through `luac` for the correct minor version of Lua, in
case newer versions would allow syntax that isn't backwards compatible.