Solve "robotics" prerequisite problem with a new setting + 8 strings
This *is* the overengineered solution.

During testing I became very annoyed to realize that there is no support
for a `__TECHNOLOGY__name__` placeholder in localized strings, even
though `__ENTITY__name__` and `__ITEM__name__` and some others exist.
That would be really useful for strings like the ones I added that
reference the names of research technologies (or indeed, a couple of the
existing strings too).

Rich text seems to be supported in all the setting names/tooltips, but
putting an icon of Logistics Warehousing 1 isn't very useful IMO.