Solve the "robotics" technology problem by requiring both its children
Warehouses are powerful enough, it's probably reasonable to make users
spend a couple hundred extra science packs to get this functionality.

It's also technically impossible to require Construction OR Logistic
robot research, because of how Factorio's data model works. Choosing
just one of the two would be purely arbitrary.

(So is choosing to require both, to be fair, but we have very few
options available without making things all weird with Lua scripting.)