Replace storehouse-linked sprite with recolor of storehouse-buffer
We don't have ready access to the graphics project used to make these
sprites, unfortunately. This should still be good enough; I zoomed in
really far and chose which pixels to select (or not) around the edges of
the green areas (in the source storehouse-buffer graphic) manually.

Removing the chroma component and lightening the selection got me pretty
close to how the color looks in the storehouse-linked sprite that's
being replaced (because its proportions are off).